src/data_cleaner/pyproject.toml
[project]
name = "cloud-brocade-cleaner"
version = "0.1.0"
description = "云锦图片数据清洗工具，用于SD1.5微调训练数据准备"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Dev", email = "dev@example.com"}
]

dependencies = [
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.35.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
# GPU 支持（CUDA 12.1）
gpu = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]

[project.scripts]
clean-brocade = "data_cleaner.data_cleaner:main"
run-cleaner = "data_cleaner.run_cleaner:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["data_cleaner*"]
